Separated theme and design settings
