+++ app/routes/settings/theme.js
@@ -15,11 +15,13 @@ export default AuthenticatedRoute.extend(CurrentUserSettings, {
 
     model() {
         return RSVP.hash({
-            settings: this.settings.reload()
+            settings: this.settings.reload(),
+            themes: this.store.findAll('theme')
         });
     },
 
-    setupController() {
+    setupController(controller) {
+        controller.set('themes', this.store.peekAll('theme'));
         this.controller.send('reset');
     },
 
@@ -30,29 +32,14 @@ export default AuthenticatedRoute.extend(CurrentUserSettings, {
     },
 
     actions: {
-        save() {
-            // since shortcuts are run on the route, we have to signal to the components
-            // on the page that we're about to save.
-            $('.page-actions .gh-btn-blue').focus();
-
-            this.controller.send('save');
-        },
-
-        willTransition(transition) {
-            let controller = this.controller;
-            let modelIsDirty = controller.dirtyAttributes;
-
-            if (modelIsDirty) {
-                transition.abort();
-                controller.send('toggleLeaveSettingsModal', transition);
-                return;
-            }
+        activateTheme(theme) {
+            return this.controller.send('activateTheme', theme);
         }
     },
 
     buildRouteInfoMetadata() {
         return {
-            titleToken: 'Settings - Design'
+            titleToken: 'Settings - Theme'
         };
     }
 });
