+++ app/templates/settings/theme/uploadtheme.hbs
@@ -1,0 +1,7 @@
+<GhFullscreenModal @modal="upload-theme"
+    @model={{hash
+        themes=themes
+        activate=(route-action 'activateTheme')
+    }}
+    @close={{route-action "cancel"}}
+    @modifier="action wide" />
